+++ src/edit/objects/sprite.ts
@@ -1,16 +1,20 @@
-import { Sprite as ThreeSprite, SpriteMaterial, Texture, Object3D, LessDepth, MeshBasicMaterial, PlaneGeometry, Mesh } from 'three';
+import { Sprite as ThreeSprite, SpriteMaterial, Texture, Object3D, MeshBasicMaterial, PlaneGeometry, Mesh } from 'three';
 import { BaseObject } from './baseObject';
 import { SceneObjectType } from '../sceneManager/interface';
 import SceneManager from '../sceneManager/sceneManager';
+import MaterialManager from '../materialManager';
 export interface ISpriteConfig {
     color?: string;
     opacity?: number,
     texture?: Texture
 }
 
+const SpriteOutlineMaterial = 'SpriteOutlineMaterial';
+
 export class Sprite extends BaseObject {
     public node: Object3D;
     public config: ISpriteConfig;
+    
     constructor(config: ISpriteConfig) {
         super();
         this.config = config;
@@ -25,40 +29,30 @@ export class Sprite extends BaseObject {
             ...params,
         });
         this.node = new ThreeSprite(material);        
-    
-        // TODO 后续待完善
-        const lineMaterial = new SpriteMaterial({
-            color: '#f00',
-            // 处理透明模式 不会覆盖
-            depthFunc: LessDepth,
-        })
-        const lineSprite = new ThreeSprite(lineMaterial);
-        lineSprite.scale.set(1.04, 1.04, 1.04);
-        lineSprite.renderOrder = 2; // 后渲染
-        lineSprite.visible = false;
-        this.node.add(lineSprite);
-        this.node.userData.outline = lineSprite;
 
-        const mesh = this.initHelperMesh();
-
-        this.node.userData.outline = mesh;
+        this.initOutlineMesh();
         this.connectObject();
     }
 
-    initHelperMesh() {
-        const material = new MeshBasicMaterial({
-            color: '#f00',
-            depthFunc: LessDepth,
-            transparent: true,
-        });
-        const geometry = new PlaneGeometry(1.04, 1.04);
-        const mesh = new Mesh(geometry, material);       
-        mesh.visible = false;
-        // mesh.renderOrder = 2;
-        mesh.onBeforeRender = () => {
-            this.node.rotation.copy(SceneManager.cameraManager.instance.rotation);
+    initOutlineMesh() {
+        if(!MaterialManager.hasMaterial(SpriteOutlineMaterial)) {
+            MaterialManager.setMaterial(SpriteOutlineMaterial, new MeshBasicMaterial({
+                opacity: 0,
+                transparent: true,
+                depthWrite: false,
+            }))
         }
-        this.node.add(mesh);
-        return mesh;
+        const scale = 1.02;
+        const geometry = new PlaneGeometry(scale, scale);
+        const outline = new Mesh(geometry, MaterialManager.getMaterial(SpriteOutlineMaterial));   
+        this.outline = outline; 
+        outline.onBeforeRender = () => {
+            outline.rotation.copy(SceneManager.cameraManager.instance.rotation);
+        }
+        outline.userData.type = 'outline';
+        outline.userData.connectObject = this.node;
+        // Tip: add to scene to keep outline renderpass work
+        SceneManager.scene.add(outline);
+        this.node.userData.outline = outline;
     }
 }